Removed class that wasn't used and renamed a interface.
