+++ src/CUITe/CUITe.csproj
@@ -138,7 +138,6 @@
     <Compile Include="Controls\HtmlControls\HtmlCheckBox.cs" />
     <Compile Include="Controls\HtmlControls\HtmlComboBox.cs" />
     <Compile Include="Controls\HtmlControls\HtmlControl.cs" />
-    <Compile Include="Controls\HtmlControls\HtmlControlOfT.cs" />
     <Compile Include="Controls\HtmlControls\HtmlCustom.cs" />
     <Compile Include="Controls\HtmlControls\HtmlDiv.cs" />
     <Compile Include="Controls\HtmlControls\HtmlDocument.cs" />
@@ -172,7 +171,7 @@
     <Compile Include="Controls\HtmlControls\HtmlTextArea.cs" />
     <Compile Include="Controls\HtmlControls\HtmlUnorderedList.cs" />
     <Compile Include="Controls\HtmlControls\HtmlXml.cs" />
-    <Compile Include="Controls\HtmlControls\IHtmlControl.cs" />
+    <Compile Include="Controls\HtmlControls\IHasInnerText.cs" />
     <Compile Include="Controls\ModifierKeysLifetime.cs" />
     <Compile Include="Controls\SilverlightControls\SilverlightButton.cs" />
     <Compile Include="Controls\SilverlightControls\SilverlightCalendar.cs" />
